some light refactoring of DropTipWizardFlows org
