+++ app/src/organisms/DropTipWizardFlows/hooks/useDropTipCommands.ts
@@ -3,7 +3,7 @@ import * as React from 'react'
 import { useDeleteMaintenanceRunMutation } from '@opentrons/react-api-client'
 
 import { MANAGED_PIPETTE_ID, POSITION_AND_BLOWOUT } from '../constants'
-import { getAddressableAreaFromConfig } from '../getAddressableAreaFromConfig'
+import { getAddressableAreaFromConfig } from '../utils'
 import { useNotifyDeckConfigurationQuery } from '../../../resources/deck_configuration'
 import type {
   CreateCommand,
