+++ app/src/organisms/DropTipWizardFlows/hooks/useTipAttachmentStatus/index.ts
@@ -1,70 +1,14 @@
 import * as React from 'react'
 import head from 'lodash/head'
 
+import { useInstrumentsQuery } from '@opentrons/react-api-client'
 import { getPipetteModelSpecs } from '@opentrons/shared-data'
 
 import { getPipettesWithTipAttached } from './getPipettesWithTipAttached'
-import { useDropTipRouting, useDropTipWithType } from './hooks'
-import { DropTipWizard } from './DropTipWizard'
 
-import type { PipetteModelSpecs, RobotType } from '@opentrons/shared-data'
-import type { Mount, PipetteData } from '@opentrons/api-client'
-import type { FixitCommandTypeUtils, IssuedCommandsType } from './types'
+import type { Mount } from '@opentrons/api-client'
+import type { PipetteModelSpecs } from '@opentrons/shared-data'
 import type { GetPipettesWithTipAttached } from './getPipettesWithTipAttached'
-import { useInstrumentsQuery } from '@opentrons/react-api-client'
-
-/** Provides the user toggle for rendering Drop Tip Wizard Flows.
- *
- * NOTE: Rendering these flows is independent of whether tips are actually attached. First use useTipAttachmentStatus
- * to get tip attachment status.
- */
-export function useDropTipWizardFlows(): {
-  showDTWiz: boolean
-  toggleDTWiz: () => void
-} {
-  const [showDTWiz, setShowDTWiz] = React.useState(false)
-
-  const toggleDTWiz = (): void => {
-    setShowDTWiz(!showDTWiz)
-  }
-
-  return { showDTWiz, toggleDTWiz }
-}
-
-export interface DropTipWizardFlowsProps {
-  robotType: RobotType
-  mount: PipetteData['mount']
-  instrumentModelSpecs: PipetteModelSpecs
-  /* isTakeover allows for optionally specifying a different callback if a different client cancels the "setup" type flow. */
-  closeFlow: (isTakeover?: boolean) => void
-  /* Optional. If provided, DT will issue "fixit" commands and render alternate Error Recovery compatible views. */
-  fixitCommandTypeUtils?: FixitCommandTypeUtils
-}
-
-export function DropTipWizardFlows(
-  props: DropTipWizardFlowsProps
-): JSX.Element {
-  const { fixitCommandTypeUtils } = props
-
-  const issuedCommandsType: IssuedCommandsType =
-    fixitCommandTypeUtils != null ? 'fixit' : 'setup'
-
-  const dropTipWithTypeUtils = useDropTipWithType({
-    ...props,
-    issuedCommandsType,
-  })
-
-  const dropTipRoutingUtils = useDropTipRouting(fixitCommandTypeUtils)
-
-  return (
-    <DropTipWizard
-      {...props}
-      {...dropTipWithTypeUtils}
-      {...dropTipRoutingUtils}
-      issuedCommandsType={issuedCommandsType}
-    />
-  )
-}
 
 const INSTRUMENTS_POLL_MS = 5000
