+++ app/src/organisms/DropTipWizardFlows/index.ts
@@ -1,0 +1,2 @@
+export * from './DropTipWizardFlows'
+export { useTipAttachmentStatus, useHomePipettes } from './hooks'
