+++ app/src/organisms/DropTipWizardFlows/utils/getAddressableAreaFromConfig.ts
