+++ app/src/organisms/DropTipWizardFlows/utils/index.ts
@@ -1,0 +1,1 @@
+export * from './getAddressableAreaFromConfig'
